feat: add weather card
