+++ src/assets/styles/constants/theme.css
@@ -12,6 +12,7 @@
   --primary-bg-color: var(--lighten-blue);
   --secondary-bg-color: var(--blue);
   --chart-bg-color: var(--lighten-blue);
+  --card-bg-color: var(--white);
 }
 
 .dark-theme {
@@ -20,4 +21,5 @@
   --primary-bg-color: var(--black);
   --secondary-bg-color: var(--darken-gray);
   --chart-bg-color: var(--darken-gray);
+  --card-bg-color: var(--darken-gray);
 }
