+++ src/components/WeatherCard/WeatherCard.module.scss
@@ -1,0 +1,34 @@
+@import '../../assets/styles/constants/main.scss';
+
+.weather-card {
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  min-width: 120px;
+  height: 100%;
+  background-color: $card-bg-color;
+  margin-right: 20px;
+  padding: 10px;
+
+  .date {
+    font-size: $font-size-xs;
+  }
+
+  .day-of-week {
+    font-size: $font-size-sm;
+  }
+
+  .weather-icon {
+    width: 50%;
+    margin: 10px 0;
+  }
+
+  .added-weather-info {
+    display: flex;
+
+    .temperature {
+      font-size: $font-size-sm;
+    }
+  }
+}
