+++ src/layouts/CityForecast/CityForecast.module.scss
@@ -35,7 +35,7 @@
     display: flex;
     justify-content: center;
     align-items: baseline;
-    margin: 15px 0;
+    margin: 10px 0;
 
     .temperature-degrees {
       font-size: $font-size-xl;
@@ -56,6 +56,10 @@
     display: block;
     text-align: center;
     font-size: $font-size-md;
+
+    .location {
+      margin-left: 5px;
+    }
   }
 
   .added-info {
@@ -67,7 +71,7 @@
     .added-info-value {
       display: flex;
       flex-direction: column;
-      margin: 0 15px;
+      margin: 0 10px;
     }
 
     .added-info-type-name {
