+++ src/layouts/DetailedForecast/DailyForecast/DailyForecast.module.scss
@@ -1,0 +1,12 @@
+@import '../../../assets/styles/constants/main.scss';
+
+.daily-forecast {
+  box-sizing: border-box;
+  display: flex;
+  height: 25%;
+
+  .daily-forecast-title {
+    font-size: $font-size-md;
+    margin-bottom: 10px;
+  }
+}
